requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
aiofiles==0.7.0
asgiref==3.4.1; python_version >= '3.6'
certifi==2021.5.30
charset-normalizer==2.0.3; python_version >= '3'
click==8.0.1; python_version >= '3.6'
cycler==0.10.0
decorator==4.4.2
fastapi==0.67.0
h11==0.12.0; python_version >= '3.6'
idna==3.2; python_version >= '3'
imageio==2.9.0
importlib-metadata==4.6.1; python_version < '3.8'
kiwisolver==1.3.1; python_version >= '3.6'
matplotlib==3.4.2; python_version >= '3.7'
networkx==2.5.1; python_version >= '3.6'
numpy==1.21.1
pillow==8.3.1
pydantic==1.8.2; python_full_version >= '3.6.1'
pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-multipart==0.0.5
pywavelets==1.1.1; python_version >= '3.5'
requests==2.26.0
scikit-image==0.18.2
scipy==1.7.0; python_version < '3.10' and python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
starlette==0.14.2; python_version >= '3.6'
tifffile==2021.7.2; python_version >= '3.7'
torch==1.9.0; python_full_version >= '3.6.2'
torchvision==0.10.0
typing-extensions==3.10.0.0; python_version < '3.8'
urllib3==1.26.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
uvicorn==0.14.0
zipp==3.5.0; python_version >= '3.6'
